+++ loan_calculator/schedule/base.py
@@ -1,4 +1,11 @@
-import numpy as np
+from enum import Enum
+
+
+class AmortizationScheduleType(Enum):
+
+    progressive_price_schedule = 'progressive-price-schedule'
+    regressive_price_schedule = 'regressive-price-schedule'
+    constant_amortization_schedule = 'constant-amortization-schedule'
 
 
 class BaseSchedule(object):
@@ -26,6 +33,8 @@ class BaseSchedule(object):
         was granted until the payments' due dates.
     """
 
+    schedule_type = None
+
     def __init__(self, principal, daily_interest_rate, return_days):
         """Initialize schedule."""
 
@@ -34,19 +43,19 @@ class BaseSchedule(object):
         self.return_days = return_days
 
         self.balance = getattr(
-            self, 'calculate_balance', np.zeros(len(return_days) + 1)
+            self, 'calculate_balance', (len(return_days) + 1) * [0]
         )()
 
         self.due_payments = getattr(
-            self, 'calculate_due_payments', np.zeros(len(return_days))
+            self, 'calculate_due_payments', len(return_days) * [0]
         )()
 
         self.interest_payments = getattr(
-            self, 'calculate_interest', np.zeros(len(return_days))
+            self, 'calculate_interest', len(return_days) * [0]
         )()
 
         self.amortizations = getattr(
-            self, 'calculate_amortizations', np.zeros(len(return_days))
+            self, 'calculate_amortizations', len(return_days) * [0]
         )()
 
     def calculate_due_payments(self):
@@ -63,12 +72,12 @@ class BaseSchedule(object):
 
     @property
     def total_paid(self):
-        return np.sum(getattr(self, 'due_payments', 0.0))
+        return sum(getattr(self, 'due_payments', 0.0))
 
     @property
     def total_amortization(self):
-        return np.sum(getattr(self, 'amortizations', 0.0))
+        return sum(getattr(self, 'amortizations', 0.0))
 
     @property
     def total_interest(self):
-        return np.sum(getattr(self, 'interest_payments', 0.0))
+        return sum(getattr(self, 'interest_payments', 0.0))