+++ tests/grossup/test_grossup_functions.py
@@ -1,4 +1,5 @@
-from numpy.testing import assert_almost_equal
+import pytest
+
 from loan_calculator.grossup.functions import (
     br_iof_regressive_price_grossup,
     br_iof_progressive_price_grossup,
@@ -14,7 +15,7 @@ def test_br_progressive_price_grossup_basic_evaluation():
                                            [1, 2],
                                            0.4)
 
-    assert_almost_equal(gup, 2.0, decimal=2)
+    assert gup == pytest.approx(2.0, rel=0.01)
 
 
 def test_br_regressive_price_grossup_basic_evaluation():
@@ -25,7 +26,7 @@ def test_br_regressive_price_grossup_basic_evaluation():
                                           [1, 2],
                                           0.4)
 
-    assert_almost_equal(gup, 2.0, decimal=2)
+    assert gup == pytest.approx(2.0, rel=0.01)
 
 
 def test_br_constant_amortization_grossup_basic_evaluation():
@@ -36,4 +37,4 @@ def test_br_constant_amortization_grossup_basic_evaluation():
                                                [1, 2],
                                                0.49)
 
-    assert_almost_equal(gup, 2.0, decimal=2)
+    assert gup == pytest.approx(2.0, rel=0.01)