+++ tests/schedule/test_price_schedules.py
@@ -1,6 +1,4 @@
-import numpy as np
-from numpy import testing as npt
-
+import pytest
 from loan_calculator.schedule.price import (ProgressivePriceSchedule,
                                             RegressivePriceSchedule)
 
@@ -17,41 +15,33 @@ def test_progressive_price_amortization_schedule():
         return_days
     )
 
-    expected_amortizations = np.array(
-        [744.09, 766.42, 789.41, 813.09, 837.48,
-         862.61, 888.49, 915.14, 942.60, 970.87],
-        dtype=float
-    )
+    expected_amortizations = [
+        744.09, 766.42, 789.41, 813.09, 837.48,
+        862.61, 888.49, 915.14, 942.60, 970.87
+    ]
 
-    expected_balances = np.array(
-        [8530.20, 7786.11, 7019.69, 6230.28, 5417.19,
-         4579.71, 3717.10, 2828.61, 1913.47, 970.87, 0.0],
-        dtype=float
-    )
+    expected_balances = [
+        8530.20, 7786.11, 7019.69, 6230.28, 5417.19,
+        4579.71, 3717.10, 2828.61, 1913.47, 970.87, 0.0
+    ]
 
-    expected_interests = np.array(
-        [255.91, 233.58, 210.59, 186.91, 162.52,
-         137.39, 111.51, 84.86, 57.40, 29.13],
-        dtype=float
-    )
+    expected_interests = [
+        255.91, 233.58, 210.59, 186.91, 162.52,
+        137.39, 111.51, 84.86, 57.40, 29.13
+    ]
 
-    expected_payments = np.array(
-        [1000.00, 1000.00, 1000.00, 1000.00, 1000.00,
-         1000.00, 1000.00, 1000.00, 1000.00, 1000.00],
-        dtype=float
-    )
+    expected_payments = [
+        1000.00, 1000.00, 1000.00, 1000.00, 1000.00,
+        1000.00, 1000.00, 1000.00, 1000.00, 1000.00
+    ]
 
-    npt.assert_almost_equal(schedule.amortizations, expected_amortizations,
-                            decimal=2)
-    npt.assert_almost_equal(schedule.balance, expected_balances, decimal=2)
-    npt.assert_almost_equal(schedule.interest_payments, expected_interests,
-                            decimal=2)
-    npt.assert_almost_equal(schedule.due_payments, expected_payments,
-                            decimal=2)
-
-    npt.assert_almost_equal(schedule.total_amortization, principal, decimal=2)
-    npt.assert_almost_equal(schedule.total_interest, 1469.80, decimal=2)
-    npt.assert_almost_equal(schedule.total_paid, 10000.00, decimal=2)
+    assert schedule.amortizations == pytest.approx(expected_amortizations, rel=0.01)  # noqa
+    assert schedule.balance == pytest.approx(expected_balances, rel=0.01)
+    assert schedule.interest_payments == pytest.approx(expected_interests, rel=0.01)  # noqa
+    assert schedule.due_payments == pytest.approx(expected_payments, rel=0.01)
+    assert schedule.total_amortization == pytest.approx(principal, rel=0.01)
+    assert schedule.total_interest == pytest.approx(1469.80, rel=0.01)
+    assert schedule.total_paid == pytest.approx(10000.00, rel=0.01)
 
 
 def test_regressive_price_amortization_schedule():
@@ -66,38 +56,30 @@ def test_regressive_price_amortization_schedule():
         return_days
     )
 
-    expected_amortizations = np.array(
-        [970.87, 942.6, 915.14, 888.49, 862.61,
-         837.48, 813.09, 789.41, 766.42, 744.09],
-        dtype=float
-    )
+    expected_amortizations = [
+        970.87, 942.6, 915.14, 888.49, 862.61,
+        837.48, 813.09, 789.41, 766.42, 744.09
+    ]
 
-    expected_balances = np.array(
-        [8530.20, 7786.11, 7019.69, 6230.28, 5417.19,
-         4579.71, 3717.10, 2828.61, 1913.47, 970.87, 0.0],
-        dtype=float
-    )
+    expected_balances = [
+        8530.20, 7786.11, 7019.69, 6230.28, 5417.19,
+        4579.71, 3717.10, 2828.61, 1913.47, 970.87, 0.0
+    ]
 
-    expected_interests = np.array(
-        [29.13, 57.4, 84.86, 111.51, 137.39,
-         162.52, 186.91, 210.59, 233.58, 255.91],
-        dtype=float
-    )
+    expected_interests = [
+        29.13, 57.4, 84.86, 111.51, 137.39,
+        162.52, 186.91, 210.59, 233.58, 255.91
+    ]
 
-    expected_payments = np.array(
-        [1000.00, 1000.00, 1000.00, 1000.00, 1000.00,
-         1000.00, 1000.00, 1000.00, 1000.00, 1000.00],
-        dtype=float
-    )
+    expected_payments = [
+        1000.00, 1000.00, 1000.00, 1000.00, 1000.00,
+        1000.00, 1000.00, 1000.00, 1000.00, 1000.00
+    ]
 
-    npt.assert_almost_equal(schedule.amortizations, expected_amortizations,
-                            decimal=2)
-    npt.assert_almost_equal(schedule.balance, expected_balances, decimal=2)
-    npt.assert_almost_equal(schedule.interest_payments, expected_interests,
-                            decimal=2)
-    npt.assert_almost_equal(schedule.due_payments, expected_payments,
-                            decimal=2)
-
-    npt.assert_almost_equal(schedule.total_amortization, principal, decimal=2)
-    npt.assert_almost_equal(schedule.total_interest, 1469.80, decimal=2)
-    npt.assert_almost_equal(schedule.total_paid, 10000.00, decimal=2)
+    assert schedule.amortizations == pytest.approx(expected_amortizations, rel=0.01)  # noqa
+    assert schedule.balance == pytest.approx(expected_balances, rel=0.01)
+    assert schedule.interest_payments == pytest.approx(expected_interests, rel=0.01)  # noqa
+    assert schedule.due_payments == pytest.approx(expected_payments, rel=0.01)
+    assert schedule.total_amortization == pytest.approx(principal, rel=0.01)
+    assert schedule.total_interest == pytest.approx(1469.80, rel=0.01)
+    assert schedule.total_paid == pytest.approx(10000.00, rel=0.01)