+++ tests/test_loan.py
@@ -1,10 +1,9 @@
+import pytest
+
 from datetime import date
 
-import pytest
-from numpy.testing import assert_almost_equal
-
 from loan_calculator.loan import Loan
-
+from loan_calculator.schedule.base import AmortizationScheduleType
 
 args_ = (
     1000.0,
@@ -15,50 +14,6 @@ args_ = (
         date(2020, 1, 4), date(2020, 1, 5),
     ]
 )
-
-
-def test_very_large_year_size():
-    """Assert trivial year provides equal amortizations."""
-
-    def _do_assert(schedule):
-        loan = Loan(
-            *args_,
-            year_size=100000,
-            # grace_period=0,
-            amortization_schedule_type=schedule
-        )
-
-        assert_almost_equal(
-            [250.0, 250.0, 250.0, 250.0],
-            loan.amortizations,
-            2,
-            'Unexpected amortizations for schedule %s'
-            % schedule,
-        )
-
-        assert_almost_equal(
-            [0.0, 0.0, 0.0, 0.0],
-            loan.interest_payments,
-            2,
-            'Unexpected interests for schedule %s'
-            % schedule
-        )
-
-        assert_almost_equal(
-            loan.due_payments,
-            loan.amortizations,
-            2,
-            'Unexpected payments/amortizations for schedule %s'
-            % schedule
-        )
-
-    for schedule in [
-        'progressive_price_schedule',
-        'regressive_price_schedule',
-        'constant_amortization_schedule',
-    ]:
-
-        _do_assert(schedule)
 
 
 def test_grace_period_exceeds_loan_start():
@@ -75,6 +30,6 @@ def test_grace_period_exceeds_loan_start():
 
 def test_unknown_amortization_schedule():
 
-    with pytest.raises(TypeError):
+    with pytest.raises(ValueError):
 
         Loan(*args_, amortization_schedule_type='foo')