+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+1.2.0 (2020-12-17)
+------------------
+
+* Implement Newton-Raphson based solver to approximate IRR
+* Remove dependency on `scipy` and `numpy`
+* Expand on grossup documentation
+* Implement function `interest_rate.convert_to_daily_interest_rate`
+
 1.1.0 (2020-10-15)
 ------------------